Create the function that performs the operations
Create three variables: one for the first operand, one for the operator
and one for the second operand. They will be used later to update the
calculator display.

Fix a bug in the add and subtract functions: The initial value
of the reduce methods was wrong.

Update the nomenclature of the parameters of ths operation functions.