Create the functions that populate the display
Create the functions that make the number and operator buttons appear
on the screen when you click on them. Also the function that clears the
screen.

Create the variable that will contain the display value.

Move global variables to the top of the "script.js" file.

Remove screen content that was hardcoded in the "index.hmtl" file.