Create the "equals" function
Modify the function of the operator buttons to store the first operand
and the operator of the current operation.

The function of the equals button will store the second operand,
perform the operation and display the result on the screen.

Modify the multiply operation case of the switch in the operate
function to match the case character with the button character.