Refactor: eval one pair of numbers at a time.
Give global variables an initial empty or 0 state.

Give a default case to the switch conditional of the "operate" function
to return 0.

Move the "clear" and "equals" functions aside and refactor their
contents to perform the desired function.

Refactor the listeners of the number, operation and decimal point
buttons to evaluate a single pair of numbers at a time.